+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/impl/CertifiedProductSearchManagerTest.java
@@ -43,7 +43,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		SearchRequest searchRequest = new SearchRequest();
 		searchRequest.setDeveloper("Test Developer 1");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(2, response.getResults().size());
+		assertEquals(3, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getDeveloper().get("name").toString().startsWith("Test Developer 1"));
@@ -57,7 +57,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		SearchRequest searchRequest = new SearchRequest();
 		searchRequest.setProduct("Test Product 1");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(2, response.getResults().size());
+		assertEquals(3, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getProduct().get("name").toString().startsWith("Test Product 1"));
@@ -100,7 +100,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		searchRequest.setCertificationBody("InfoGard");
 		searchRequest.setVisibleOnCHPL("BOTH");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(4, response.getResults().size());
+		assertEquals(5, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getCertifyingBody().get("name").toString().startsWith("InfoGard"));
@@ -130,7 +130,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		searchRequest.setPracticeType("Ambulatory");
 		searchRequest.setVisibleOnCHPL("BOTH");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(4, response.getResults().size());
+		assertEquals(5, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getPracticeType().get("name").toString().startsWith("Ambulatory"));
@@ -145,7 +145,7 @@ public class CertifiedProductSearchManagerTest extends TestCase {
 		SearchRequest searchRequest = new SearchRequest();
 		searchRequest.setVisibleOnCHPL("YES");
 		SearchResponse response = certifiedProductSearchManager.search(searchRequest);
-		assertEquals(2, response.getResults().size());
+		assertEquals(3, response.getResults().size());
 		
 		for (CertifiedProductSearchResult result : response.getResults() ){
 			assertTrue(result.getVisibleOnChpl());