+++ RELEASE_NOTES.md
@@ -1,5 +1,19 @@
 # Release Notes
 
+## Version 1.7.0
+_25 july 2016_
+
+### Features Added
+* Add validation to CHPL product update so that it matches validation done on CHPL product confirm
+* Pass errors in test tools and test functionality back to the UI
+* Add task rating standard deviation to upload, confirm, and api get/update
+
+### Bugs Fixed
+* Check test functionality for invalid values on product confirm and edit
+* Check test tools for invalid values on product confirm and edit
+
+---
+
 ## Version 1.6.1
 _1 July 2016_